Split Simple Model into own repo
