+++ data_pipeline_api/ext/SEIRSconfigWindows.yaml
