+++ data_pipeline_api/ext/static_params_SEIRS.csv
