--- simpleModel/__init__.py
@@ -1,15 +1,0 @@
-__all__ = [
-    "SEIRS_Model",
-    "SEIRS_Plot",
-    "write_model_to_csv",
-    "getResourceDirectory",
-    "readInitialParameters",
-]
-
-from .common.SEIRS_Model import (
-    SEIRS_Model,
-    getResourceDirectory,
-    readInitialParameters,
-    write_model_to_csv,
-)
-from .common.SEIRS_Plot import SEIRS_Plot
